feat: Inversion of control for how injecting config works in deploy workflow
BREAKING CHANGE: Instead of passing a boolean input called "apply_config" that
would cause the deploy workflow to run a predefined script in package.json, we
now allow the workflow consumer to pass the entire command, including the
paramaters in order and fashion that fits the usecase best.
This allows to pass additional arguments to those scripts, not related to
cursor deployment, without the deploy workflow needing to support them explicitly.