feat: add environment suffix to github artifact
currently we're only using the tree_hash as name of our artifact, but tree_hash is the same across all environments. So we're overwriting existing artifacts and only have a single one for production in the end.
After this change we should have an artifact per environment.
I'm hopeful that this will also fix the errors we're seeing since introducing the product-dev environment, my assumption is that github locks the artifact for a short time during upload and when the stars align product-dev and staging upload at the same time, creating this error.